修改项目中所有的snapshotDate为lastMonth，本系统目前只算当前月的上个月的数据

+++ src/main/java/com/perfectchina/bns/repositories/CustomerBonusRateRepository.java
@@ -21,7 +21,7 @@ public interface CustomerBonusRateRepository extends JpaRepository<CustomerBonus
 	//find all rate >= lastMonth rate base on LMAOPV asc
 	@Query("SELECT br FROM CustomerBonusRate br WHERE br.aopv > :aopv and ( :checkAsAtDate between br.effectiveFrom and br.effectiveTo ) order by br.aopv asc ")
 	List<CustomerBonusRate> findBonusRateByAopvAndDateAsc(
-			@Param("aopv") long aopv,
+			@Param("aopv") float aopv,
 			@Param("checkAsAtDate") Date checkAsAtDate);
 
 }
